# Conflicts:
#	packages/page-contracts/src/index.tsx

+++ packages/apps-config/src/api/spec/encointer-node-teeproxy.ts
@@ -42,5 +42,5 @@ export default {
     url: 'Text'
   },
   // weight changed to u64 since 2.0.0-rc1 (commit 2051ecbf79e April 16th 2020
-  Weight: 'u32'
+  Weight: 'u64'
 };
